refactor: stack backend-only sin frontend para integración con transaccional ADUACOL
- Nuevo stack AduacolBackendStack: solo pipeline de procesamiento (sin Cognito, AppSync, CloudFront, CodeBuild)
- Elimina 1 VPC Endpoint (CODEBUILD) = ~$7/mes menos
- Outputs exponen bucket names + ARNs + KMS key para que ADUACOL configure su IAM
- .gitignore actualizado: excluye samples de ejemplo, config libraries de demo, frontend files, .DS_Store
- main.ts apunta al nuevo stack
- bedrock-llm-stack.ts y optional-admin.ts removidos del tracking (conservados localmente como referencia)
